refactor: move auth options to separate file and update imports
